bunch of improvements
